requirements.txt
pyspark==3.5.1
kafka-python==2.0.2
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
pyarrow==16.1.0
streamlit==1.37.1
plotly==5.23.0
python-dotenv==1.0.1
joblib==1.4.2
psutil==6.0.0
requests==2.32.3
pytest==8.3.2
